alx_travel_app/requirements.txt
Django>=3.2
 djangorestframework
drf-yasg
